+++ uliweb/core/js.py
@@ -1,6 +1,7 @@
 import re
 import datetime
 import decimal
+from html import Builder
 
 ESCAPE = re.compile(r'[\x00-\x1f\\"\b\f\n\r\t]')
 ESCAPE_DCT = {
@@ -106,6 +107,8 @@ class JSONEncoder(object):
             yield '"' + obj.strftime('%Y-%m-%d') + '"'
         elif isinstance(obj, datetime.time):
             yield '"' + obj.strftime('%H:%M:%S') + '"'
+        elif isinstance(obj, Builder):
+            yield str(obj)
         else:
             raise TypeError(repr(obj) + " can't be serialized to JSON")