+++ uliweb/template_files/app/views.py
@@ -1,5 +1,5 @@
 #coding=utf-8
-from uliweb import expose
+from uliweb import expose, functions
 
 @expose('/')
 def index():