+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Uliweb Change Log
 * Improve generic command can process `camel_view` to `CamelView`
 * Add `pretty_dict` function to common
 * Fix xmlrpc not compatiable with `prepare_request`
+* Add `Builder` support in json_dumps
 
 0.1.6 Version
 -----------------